+++ src/base/EntryTypeProcessor.php
@@ -76,4 +76,53 @@ class EntryTypeProcessor extends Processor
     {
         return Craft::$app->sections->saveEntryType($item);
     }
+
+    /**
+     * @param $item
+     * @param array $extraAttributes
+     *
+     * @return array
+     *
+     * @throws \yii\base\InvalidConfigException
+     */
+    public function export($item, array $extraAttributes = [])
+    {
+        /** @var EntryType $item */
+        $attributeObj = [];
+        $extraAttributes = array_merge($extraAttributes, $this->additionalAttributes(get_class($item)));
+        foreach($extraAttributes as $attribute) {
+            $attributeObj[$attribute] = $item->$attribute;
+        }
+        $hasTitleField = boolval($item->hasTitleField);
+        $entryTypeObj = array_merge([
+            'sectionHandle' => $item->getSection()->handle,
+            'name' => $item->name,
+            'handle' => $item->handle,
+            'hasTitleField' => $hasTitleField,
+            'titleLabel' => ($hasTitleField) ? $item->titleLabel : "",
+            'titleFormat' => (!$hasTitleField) ? $item->titleFormat : "",
+            'fieldLayout' => $this->exportFieldLayout($item->getFieldLayout()),
+            'requiredFields' => $this->exportRequiredFields($item->getFieldLayout()),
+        ], $attributeObj);
+        
+        if (count($entryTypeObj['requiredFields']) <= 0) {
+            unset($entryTypeObj['requiredFields']);
+        }
+
+        return $this->stripNulls($entryTypeObj);
+    }
+
+    /**
+     * @param $id
+     *
+     * @return array
+     *
+     * @throws \yii\base\InvalidConfigException
+     */
+    public function exportById($id)
+    {
+        $entryType = Craft::$app->sections->getEntryTypeById($id);
+
+        return $this->export($entryType);
+    }
 }