+++ src/base/GlobalSetProcessor.php
@@ -52,4 +52,45 @@ class GlobalSetProcessor extends Processor
     {
         return Craft::$app->globals->saveSet($item);
     }
+
+    /**
+     * @param $item
+     * @param array $extraAttributes
+     *
+     * @return array
+     */
+    public function export($item, array $extraAttributes = [])
+    {
+        /** @var GlobalSet $item */
+        $attributeObj = [];
+        $extraAttributes = array_merge($extraAttributes, $this->additionalAttributes(get_class($item)));
+        foreach($extraAttributes as $attribute) {
+            $attributeObj[$attribute] = $item->$attribute;
+        }
+
+        $globalSetObj = array_merge([
+            'name' => $item->name,
+            'handle' => $item->handle,
+            'fieldLayout' => $this->exportFieldLayout($item->getFieldLayout()),
+            'requiredFields' => $this->exportRequiredFields($item->getFieldLayout()),
+        ], $attributeObj);
+
+        if (count($globalSetObj['requiredFields']) <= 0) {
+            unset($globalSetObj['requiredFields']);
+        }
+
+        return $this->stripNulls($globalSetObj);
+    }
+
+    /**
+     * @param $id
+     *
+     * @return array
+     */
+    public function exportById($id)
+    {
+        $globalSet = Craft::$app->globals->getSetById($id);
+
+        return $this->export($globalSet);
+    }
 }