+++ src/base/SectionProcessor.php
@@ -13,6 +13,7 @@ namespace pennebaker\architect\base;
 use Craft;
 use craft\models\Section;
 use craft\models\Section_SiteSettings;
+use pennebaker\architect\Architect;
 
 /**
  * SectionProcessor defines the common interface to be implemented by plugin classes.
@@ -55,5 +56,84 @@ class SectionProcessor extends Processor
     public function save($item, bool $update = false)
     {
         return Craft::$app->sections->saveSection($item);
+    }
+
+    /**
+     * @param string $class
+     *
+     * @return array|mixed
+     */
+    public function additionalAttributes(string $class) {
+        $additionalAttributes = [
+            'structure' => [
+                'maxLevels',
+            ],
+            'channel' => [
+                'propagateEntries'
+            ],
+        ];
+        return (isset($additionalAttributes[$class])) ? $additionalAttributes[$class] : [];
+    }
+
+    /**
+     * @param $item
+     * @param array $extraAttributes
+     *
+     * @return array
+     *
+     * @throws \yii\base\InvalidConfigException
+     */
+    public function export($item, array $extraAttributes = [])
+    {
+        /** @var Section $item */
+        $attributeObj = [];
+        $extraAttributes = array_merge($extraAttributes, $this->additionalAttributes(get_class($item)), $this->additionalAttributes($item->type));
+        foreach($extraAttributes as $attribute) {
+            if ($attribute === 'propagateEntries') {
+                $attributeObj[$attribute] = boolval($item->$attribute);
+            } else {
+                $attributeObj[$attribute] = $item->$attribute;
+            }
+        }
+
+        $sectionObj = array_merge([
+            'name' => $item->name,
+            'handle' => $item->handle,
+            'type' => $item->type,
+            'enableVersioning' => boolval($item->enableVersioning),
+        ], $attributeObj);
+
+        $siteSettings = $item->getSiteSettings();
+        $sectionObj['siteSettings'] = [];
+        foreach ($siteSettings as $siteSetting) {
+            array_push($sectionObj['siteSettings'], [
+                'siteId' => ($siteSetting->getSite()->primary) ? null : $siteSetting->getSite()->handle,
+                'hasUrls' => $siteSetting->hasUrls,
+                'uriFormat' => $siteSetting->uriFormat,
+                'template' => $siteSetting->template,
+                'enabledByDefault' => boolval($siteSetting->enabledByDefault),
+            ]);
+        }
+        $entryTypes = $item->getEntryTypes();
+        $sectionObj['entryTypes'] = [];
+        foreach ($entryTypes as $entryType) {
+            array_push($sectionObj['entryTypes'], Architect::$processors->entryTypes->export($entryType));
+        }
+
+        return $this->stripNulls($sectionObj);
+    }
+
+    /**
+     * @param $id
+     *
+     * @return array
+     *
+     * @throws \yii\base\InvalidConfigException
+     */
+    public function exportById($id)
+    {
+        $section = Craft::$app->sections->getSectionById($id);
+
+        return $this->export($section);
     }
 }