+++ src/translations/en/architect.php
@@ -24,7 +24,6 @@ return [
     'Architect plugin loaded' => 'Architect plugin loaded',
     'Import' => 'Import',
     'Export' => 'Export',
-    'Migrations' => 'Migrations',
     "Import CP Section body goes here" => "Import CP Section body goes here",
     'JSON Data' => 'JSON Data',
     'Put your json import data here.' => 'Put your json import data here.',
@@ -34,5 +33,5 @@ return [
     'Possible values:' => 'Possible values:',
     'No field type matching "{fieldType}".' => 'No field type matching "{fieldType}".',
     'Field layout can only have 1 tab named "Content".' => 'Field layout can only have 1 tab named "Content".',
-
+    'Group id is invalid.' => 'Group id is invalid.',
 ];